Move mixin into own file, embed in values.yaml
